Renaming of analysis board related classes
