+++ lib/src/model/analysis/analysis_controller.dart
@@ -6,8 +6,8 @@ import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:intl/intl.dart';
 import 'package:lichess_mobile/src/model/account/account_service.dart';
-import 'package:lichess_mobile/src/model/analysis/abstract_analysis_state.dart';
 import 'package:lichess_mobile/src/model/analysis/analysis_preferences.dart';
+import 'package:lichess_mobile/src/model/analysis/common_analysis_state.dart';
 import 'package:lichess_mobile/src/model/analysis/opening_service.dart';
 import 'package:lichess_mobile/src/model/analysis/server_analysis_service.dart';
 import 'package:lichess_mobile/src/model/common/chess.dart';
@@ -649,7 +649,7 @@ class AnalysisController extends _$AnalysisController
 @freezed
 sealed class AnalysisState
     with _$AnalysisState
-    implements EvaluationMixinState, AbstractAnalysisState {
+    implements EvaluationMixinState, CommonAnalysisState {
   const AnalysisState._();
 
   const factory AnalysisState({
@@ -773,7 +773,9 @@ sealed class AnalysisState
 }
 
 @freezed
-sealed class AnalysisCurrentNode with _$AnalysisCurrentNode implements AbstractAnalysisCurrentNode {
+sealed class AnalysisCurrentNode
+    with _$AnalysisCurrentNode
+    implements AnalysisCurrentNodeInterface {
   const AnalysisCurrentNode._();
 
   const factory AnalysisCurrentNode({
