+++ lib/src/model/analysis/analysis_preferences.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:lichess_mobile/src/model/analysis/abstract_analysis_prefs.dart';
+import 'package:lichess_mobile/src/model/analysis/common_analysis_prefs.dart';
 import 'package:lichess_mobile/src/model/settings/preferences_storage.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
@@ -58,7 +58,7 @@ class AnalysisPreferences extends _$AnalysisPreferences with PreferencesStorage<
 }
 
 @Freezed(fromJson: true, toJson: true)
-sealed class AnalysisPrefs with _$AnalysisPrefs implements Serializable, AbstractAnalysisPrefs {
+sealed class AnalysisPrefs with _$AnalysisPrefs implements Serializable, CommonAnalysisPrefs {
   const AnalysisPrefs._();
 
   const factory AnalysisPrefs({
