+++ lib/src/model/broadcast/broadcast_analysis_controller.dart
@@ -5,8 +5,8 @@ import 'package:deep_pick/deep_pick.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/widgets.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:lichess_mobile/src/model/analysis/abstract_analysis_state.dart';
 import 'package:lichess_mobile/src/model/analysis/analysis_controller.dart';
+import 'package:lichess_mobile/src/model/analysis/common_analysis_state.dart';
 import 'package:lichess_mobile/src/model/broadcast/broadcast_preferences.dart';
 import 'package:lichess_mobile/src/model/broadcast/broadcast_repository.dart';
 import 'package:lichess_mobile/src/model/common/chess.dart';
@@ -509,7 +509,7 @@ class BroadcastAnalysisController extends _$BroadcastAnalysisController
 @freezed
 sealed class BroadcastAnalysisState
     with _$BroadcastAnalysisState
-    implements EvaluationMixinState, AbstractAnalysisState {
+    implements EvaluationMixinState, CommonAnalysisState {
   const BroadcastAnalysisState._();
 
   const factory BroadcastAnalysisState({
