+++ lib/src/model/broadcast/broadcast_preferences.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:lichess_mobile/src/model/analysis/abstract_analysis_prefs.dart';
+import 'package:lichess_mobile/src/model/analysis/common_analysis_prefs.dart';
 import 'package:lichess_mobile/src/model/settings/preferences_storage.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
@@ -63,7 +63,7 @@ class BroadcastPreferences extends _$BroadcastPreferences with PreferencesStorag
 }
 
 @Freezed(fromJson: true, toJson: true)
-sealed class BroadcastPrefs with _$BroadcastPrefs implements Serializable, AbstractAnalysisPrefs {
+sealed class BroadcastPrefs with _$BroadcastPrefs implements Serializable, CommonAnalysisPrefs {
   const factory BroadcastPrefs({
     required bool showEvaluationBar,
     @JsonKey(defaultValue: true) required bool enableServerAnalysis,
